solving failing tests
