+++ gpflow/quadrature/__init__.py
@@ -1,3 +1,5 @@
 from .base import GaussianQuadrature
 from .gauss_hermite import NDDiagGHQuadrature
-from .deprecated import hermgauss, mvhermgauss, mvnquad, ndiagquad, ndiag_mc
+from .quadrature import hermgauss, mvhermgauss, mvnquad, ndiagquad, ndiag_mc
+
+from .deprecated import ndiagquad as ndiagquad_old
