+++ gpflow/quadrature/base.py
@@ -4,8 +4,7 @@ import abc
 import tensorflow as tf
 
 
-class GaussianQuadrature():
-
+class GaussianQuadrature:
     @abc.abstractmethod
     def _build_X_W(self, mean, var):
         raise NotImplementedError
@@ -20,7 +19,9 @@ class GaussianQuadrature():
     def __call__(self, fun, mean, var, *args, **kwargs):
         X, W = self._build_X_W(mean, var)
         if isinstance(fun, Iterable):
-            # Maybe this can be better implemented by concating [f1(X), f2(X), ...]
+            # Maybe this can be better implemented by stacking [f1(X), f2(X), ...] 
             # and sum-reducing all at once
+            # The problem: there is no garantee that f1(X), f2(X), ...
+            # have comaptible shapes
             return [tf.reduce_sum(f(X, *args, **kwargs) * W, axis=0) for f in fun]
         return tf.reduce_sum(fun(X, *args, **kwargs) * W, axis=0)
